Initial work on the raspberry pi.
Not yet integrated into the the rest of the build system.

So far, all this does is boot, set up a stack, and call a c routine
which repeatedly outputs "hello world" to the activity light using morse
code.

The morse code output will probably not stick around very long.