Fixed the ordering of object files in the build system.
We had been using the linker script on x86 to ensure the multiboot
header found its way to the right place. Now, the makefile ensures that
boot.S is first on the linker command line. We do the same with the rpi.
This is a bit simpler than using the linker script.