Rewrite heap allocator
I did this in an effort to make things more readable. It's broken
though, and I'm not sure it was a good idea. It's failing the test
earlier than the last one. I'm committing it to a branch so I don't lose
it, but I may or may not actually merge.