Switch APs into protected mode, for real this time.
There was some nonsense going on before, and while bochs was reporting
things were in protected mode, they really weren't. Things in this
patch:

1. Mark real mode assembly with .code16; this is important, and was the
   main source of confusion.
2. Revert to the documented procedure for entering protected mode (cli,
   ldgt, cr0, far jump), which works now that we've dealt with (1).
3. Set the default number of CPUs for qemu and bochs to 2; we don't have
   the stack allocation logic implemented, so for now we're using a
   fixed address for the AP's stack. There obviously can't be more than one
   in this case, since they'll trample each other.

There's still a problem calling printf from ap_main, but we're able to jump
to it and actually touch the stack (works with no optimization and therefore
the function prologue/epilogue). TODO.