Hook all CPUs into the scheduler.
Spawning some threads to test. Seems to work.

One thing to note is that we're completely ignoring the case where the
ready queue is empty; the demo avoids this by making sure we allocate at
least one thread per core, and that they never block (we don't really
have a way for things to block yet anyhow).